+++ packages/@leight/query/src/api/index.ts
@@ -1,13 +1,14 @@
-import {CursorSchema} from "@leight/cursor";
+import {CursorSchema}   from "@leight/cursor";
 import {
     FilterSchema,
     type IFilterSchema
-}                     from "@leight/filter";
+}                       from "@leight/filter";
+import {type IUseQuery} from "@leight/react-query";
 import {
     type ISortSchema,
     SortSchema
-}                     from "@leight/sort";
-import {z}            from "zod";
+}                       from "@leight/sort";
+import {z}              from "zod";
 
 export const ParamsSchema = z.object({});
 export type IParamsSchema = typeof ParamsSchema;
@@ -53,3 +54,5 @@ export interface IQuerySchemaProps<
     sortSchema?: TSortSchema;
     paramsSchema?: TParamsSchema;
 }
+
+export type IUseCursorCountQuery<TQuerySchema extends IQuerySchema> = IUseQuery<TQuerySchema, number>;