+++ packages/@leight/source-client/src/source/Source.tsx
@@ -1,10 +1,8 @@
 import {type IStoreProvider}    from "@leight/context-client";
+import {useCursorState}         from "@leight/cursor-client";
 import {type IQuerySchema}      from "@leight/query";
-import {type IUseQueryStore}    from "@leight/query-client";
-import {
-    type IEntitySchema,
-    type IUseQuery
-}                               from "@leight/source";
+import {type IUseQuery}         from "@leight/react-query";
+import {type IEntitySchema,}    from "@leight/source";
 import {isCallable}             from "@leight/utils";
 import {type IStoreApi}         from "@leight/zustand";
 import {
@@ -14,51 +12,59 @@ import {
 import {z}                      from "zod";
 import {type ISourceStoreProps} from "../hook";
 
-export interface ISourceProps<
+export interface ISourceInternalProps<
     TQuerySchema extends IQuerySchema,
     TSchema extends IEntitySchema,
 > {
     readonly schema: TSchema;
     readonly useQuery: IUseQuery<z.infer<TQuerySchema> | undefined, z.infer<TSchema>[]>;
-    readonly useQueryStore: IUseQueryStore<TQuerySchema>;
     readonly SourceProvider: IStoreProvider<ISourceStoreProps<TSchema>>;
     readonly children?: ((store: IStoreApi<ISourceStoreProps<TSchema>>) => ReactNode) | ReactNode;
 
     onSuccess?(entities: z.infer<TSchema>[]): void;
 }
 
-interface ISourceInternalProps<
+export type ISourceProps<
     TQuerySchema extends IQuerySchema,
     TSchema extends IEntitySchema,
-> extends Pick<ISourceProps<TQuerySchema, TSchema>, "schema" | "useQuery" | "useQueryStore" | "onSuccess" | "children"> {
+> = Omit<ISourceInternalProps<TQuerySchema, TSchema>, "schema" | "SourceProvider" | "useQuery">;
+
+interface IInternalSourceProps<
+    TQuerySchema extends IQuerySchema,
+    TSchema extends IEntitySchema,
+> extends Pick<ISourceInternalProps<TQuerySchema, TSchema>, "schema" | "useQuery" | "onSuccess" | "children"> {
     readonly sourceContext: IStoreApi<ISourceStoreProps<TSchema>>;
 }
 
-const SourceInternal = <
+const InternalSource = <
     TQuerySchema extends IQuerySchema,
     TSchema extends IEntitySchema,
 >({
       sourceContext,
       schema,
       useQuery,
-      useQueryStore,
       onSuccess,
       children,
-  }: ISourceInternalProps<TQuerySchema, TSchema>) => {
-    const {query, id: queryId} = useQueryStore(({query, id}) => ({query, id}));
-    const result               = useQuery(query, {
+  }: IInternalSourceProps<TQuerySchema, TSchema>) => {
+    const {page, size} = useCursorState(({page, size}) => ({page, size}));
+    const result       = useQuery({
+        cursor: {
+            page,
+            size,
+        }
+    }, {
         onSuccess,
     });
 
     useEffect(() => {
         if (result.isSuccess) {
-            console.log("Source update", result?.data?.[0]?.id, query.cursor);
             const $data = result.data.filter(item => schema.safeParse(item).success);
             onSuccess?.($data);
             sourceContext.state.setEntities($data);
         }
     }, [
-        queryId,
+        page,
+        size,
         result.isSuccess,
         result.isLoading,
         result.isFetching,
@@ -80,23 +86,21 @@ export const Source = <
     {
         schema,
         useQuery,
-        useQueryStore,
         onSuccess,
         SourceProvider,
         children,
         ...props
-    }: ISourceProps<TQuerySchema, TSchema>) => {
+    }: ISourceInternalProps<TQuerySchema, TSchema>) => {
     return <SourceProvider
         {...props}
     >
-        {(sourceContext) => <SourceInternal
+        {(sourceContext) => <InternalSource
             sourceContext={sourceContext}
             schema={schema}
             useQuery={useQuery}
-            useQueryStore={useQueryStore}
             onSuccess={onSuccess}
         >
             {children}
-        </SourceInternal>}
+        </InternalSource>}
     </SourceProvider>;
 };