+++ packages/@leight/source/src/api/index.ts
@@ -4,12 +4,6 @@ import {
     type IQuerySchema
 }                           from "@leight/query";
 import {type IToString}     from "@leight/utils";
-import {
-    UseMutationOptions,
-    UseMutationResult,
-    UseQueryOptions,
-    UseQueryResult
-}                           from "@tanstack/react-query";
 import {z}                  from "zod";
 
 export type ISourceName =
@@ -95,13 +89,3 @@ export namespace ISource {
         filter: TSourceSchema["Filter"];
     }
 }
-
-export type IUseQuery<
-    TRequest,
-    TResponse,
-> = (request?: TRequest, options?: UseQueryOptions<any, any, TResponse, any>) => UseQueryResult<TResponse>;
-
-export type IUseMutation<
-    TRequest,
-    TResponse,
-> = (options?: UseMutationOptions<TResponse, any, TRequest>) => UseMutationResult<TResponse, any, TRequest>;