+++ packages/@leight/import/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @leight/import
 
+## 0.5.37
+
+### Patch Changes
+
+- Version bump
+- Updated dependencies
+    - @leight/job@0.5.41
+
 ## 0.5.36
 
 ### Patch Changes