+++ packages/@leight/mantine/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @leight/mantine
 
+## 0.5.42
+
+### Patch Changes
+
+- Version bump
+- Updated dependencies
+    - @leight/file@0.5.38
+    - @leight/file-client@0.5.39
+    - @leight/i18n@0.5.38
+    - @leight/i18n-client@0.5.40
+    - @leight/layout@0.5.38
+    - @leight/utils@0.5.40
+    - @leight/utils-client@0.5.41
+
 ## 0.5.41
 
 ### Patch Changes