+++ packages/@leight/next.js-server/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @leight/next.js-server
 
+## 0.5.44
+
+### Patch Changes
+
+- Version bump
+- Updated dependencies
+    - @leight/prisma@0.5.48
+    - @leight/container@0.5.38
+    - @leight/next.js@0.5.4
+    - @leight/user@0.5.41
+    - @leight/user-server@0.5.42
+    - @leight/utils@0.5.40
+    - @leight/winston@0.5.37
+
 ## 0.5.43
 
 ### Patch Changes