+++ packages/@leight/tsconfig/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @leight/tsconfig
 
+## 0.5.40
+
+### Patch Changes
+
+- Version bump
+
 ## 0.5.39
 
 ### Patch Changes