+++ packages/@leight/turbowatch/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @leight/turbowatch
 
+## 0.5.3
+
+### Patch Changes
+
+- Version bump
+
 ## 0.5.2
 
 ### Patch Changes