+++ packages/@leight/user/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @leight/user
 
+## 0.5.41
+
+### Patch Changes
+
+- Version bump
+- Updated dependencies
+    - @leight/prisma@0.5.48
+    - @leight/source@0.5.35
+
 ## 0.5.40
 
 ### Patch Changes