+++ packages/@leight/xlsx-import/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @leight/xlsx-import
 
+## 0.5.38
+
+### Patch Changes
+
+- Version bump
+- Updated dependencies
+    - @leight/import@0.5.37
+    - @leight/job@0.5.41
+    - @leight/trpc@0.5.37
+
 ## 0.5.37
 
 ### Patch Changes